Organise files into packages according to their functions
