+++ src/main/java/shenhe/Ui.java
@@ -1,3 +1,5 @@
+package shenhe;
+
 import java.util.Scanner;
 
 public class Ui {
