+++ src/main/java/shenhe/command/Command.java
@@ -1,3 +1,8 @@
+package shenhe.command;
+import shenhe.TaskList;
+import shenhe.Ui;
+import shenhe.Storage;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws Exception;
     public abstract boolean isExit();
