+++ src/main/java/shenhe/command/DeadlineCommand.java
@@ -1,5 +1,12 @@
-import java.time.LocalDateTime;
+package shenhe.command;
 
+import java.time.LocalDateTime;
+import shenhe.TaskList;
+import shenhe.Ui;
+import shenhe.Storage;
+import shenhe.exception.EmptyTaskDescriptionException;
+import shenhe.parser.DateParser;
+import shenhe.task.Deadline;
 public class DeadlineCommand extends Command {
     private String userInput;
     public DeadlineCommand(String userInput) {
