+++ src/main/java/shenhe/command/UnmarkCommand.java
@@ -1,18 +1,25 @@
-public class MarkCommand extends Command {
+package shenhe.command;
+
+import shenhe.exception.EmptyTaskDescriptionException;
+import shenhe.TaskList;
+import shenhe.Ui;
+import shenhe.Storage;
+
+public class UnmarkCommand extends Command {
     private String userInput;
-    public MarkCommand(String userInput) {
+    public UnmarkCommand(String userInput) {
         this.userInput = userInput;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws EmptyTaskDescriptionException {
-        if (userInput.trim().length() == 4) {
+        if (userInput.trim().length() == 6) {
             throw new EmptyTaskDescriptionException();
         } else {
-            int taskNumber = Integer.parseInt(userInput.substring(4).trim());
+            int taskNumber = Integer.parseInt(userInput.substring(6).trim());
             if (taskNumber >= 1 && taskNumber <= tasks.getSize()) {
-                tasks.getTask(taskNumber - 1).markAsDone();
-                ui.showMarkMessage();
+                ui.showUnmarkMessage();
+                tasks.getTask(taskNumber - 1).markAsUndone();
                 System.out.println(tasks.getTask(taskNumber - 1).toString());
                 storage.saveTasks(tasks);
             }
