+++ src/main/java/shenhe/exception/EmptyTaskDescriptionException.java
@@ -1,3 +1,5 @@
+package shenhe.exception;
+
 public class EmptyTaskDescriptionException extends Exception {
     public EmptyTaskDescriptionException() {
         super("My dear traveller, the task description cannot be empty. Please give me something specific.");
