+++ src/main/java/shenhe/exception/InvalidListEnquiry.java
@@ -1,3 +1,5 @@
+package shenhe.exception;
+
 public class InvalidListEnquiry extends Exception {
     public InvalidListEnquiry() {
         super("Dear traveller. If you want to check out the current list, please type in 'list' only.");
