+++ src/main/java/shenhe/task/Event.java
@@ -1,3 +1,5 @@
+package shenhe.task;
+
 public class Event extends Task {
     protected String from;
     protected String to;
