+++ src/main/java/shenhe/task/Task.java
@@ -1,3 +1,5 @@
+package shenhe.task;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Objects;
